Fix tableview logic
