+++ Demo/Indication/IndicationDemoViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import HHModule
+import HHList
 import HHListExtension
 import HHIndication
